Refactor: BagScreen app build crash resolution by adding params on Add product to cart, Shifted cart redux state to checkout reducer
