+++ src/navigations/ShopitStackNavigator.js
@@ -7,12 +7,10 @@ import ProductDetailScreen from '../screens/main/ShopitStack/ProductDetailScreen
 import ShippingAddressScreen from '../screens/main/ShopitStack/CheckoutScreens/ShippingAddressScreen'
 import PaymentScreen from '../screens/main/ShopitStack/CheckoutScreens/PaymentScreen'
 import BagScreen from '../screens/main/ShopitStack/CheckoutScreens/BagScreen'
-import FiltersDrawerNavigator from './FiltersDrawerNavigator'
 import FiltersTabNavigator from './FiltersTabNavigator'
 import { Menu, ShoppingBag, Bell, Heart, Share } from '../library/icons'
 import { colors } from '../res/palette'
 import { globalStyles } from '../styles/global'
-import OrdersScreen from '../screens/main/ShopitStack/OrdersScreen'
 
 const ShopitStack = createStackNavigator()
 
@@ -83,12 +81,6 @@ function ShopitStackNavigator ({ navigation }) {
         options={{
           headerTitle: 'Payment',
           headerRight: () => <Heart size={24} style={{color: colors.black}} />
-        }}
-      />
-      <ShopitStack.Screen name="Orders" component={OrdersScreen}
-        options={{
-          headerTitle: 'Orders',
-          headerRight: ({ color, size }) => <ShoppingBag size={size} style={{color}} />
         }}
       />
     </ShopitStack.Navigator>
