+++ src/redux/reducers/checkoutReducer.js
@@ -8,7 +8,41 @@ const DEFAULT_STATE = {
     iso: null
   },
   countriesList: [],
-  paymentMethods: []
+  paymentMethods: [],
+  cart: {
+    line_items: [
+      {
+        variant: {
+          images: [
+            {
+              styles: [
+                {
+                  url: ''
+                },
+                {
+                  url: ''
+                },
+                {
+                  url: ''
+                },
+                {
+                  url: ''
+                }
+              ]
+            }
+          ],
+          option_values: [
+            {
+              presentation: ''
+            },
+            {
+              presentation: ''
+            }
+          ]
+        }
+      }
+    ]
+  }
 };
 
 let changes = null;
@@ -142,6 +176,101 @@ export default function checkoutReducer(state = DEFAULT_STATE, action) {
       return { ...state, ...changes };
 
     case 'COMPLETE_CHECKOUT_FULFILLED':
+      changes = {
+        cart: dataFormatter.deserialize(response),
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    /**
+     * GET_CART
+     */
+    case 'GET_CART_PENDING':
+      return { ...state, saving: true };
+
+    case 'GET_CART_REJECTED':
+      changes = {
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    case 'GET_CART_FULFILLED':
+      changes = {
+        cart: dataFormatter.deserialize(response),
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    /**
+     * CREATE_CART
+     */
+    case 'CREATE_CART_PENDING':
+      return { ...state, saving: true };
+
+    case 'CREATE_CART_REJECTED':
+      changes = {
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    case 'CREATE_CART_FULFILLED':
+      changes = {
+        cart: dataFormatter.deserialize(response),
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    /**
+     * ADD_ITEM
+     */
+    case 'ADD_ITEM_PENDING':
+      return { ...state, saving: true };
+
+    case 'ADD_ITEM_REJECTED':
+      changes = {
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    case 'ADD_ITEM_FULFILLED':
+      changes = {
+        cart: dataFormatter.deserialize(response),
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    /**
+     * SET_QUANTITY
+     */
+    case 'SET_QUANTITY_PENDING':
+      return { ...state, saving: true };
+
+    case 'SET_QUANTITY_REJECTED':
+      changes = {
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    case 'SET_QUANTITY_FULFILLED':
+      changes = {
+        cart: dataFormatter.deserialize(response),
+        saving: false
+      };
+      return { ...state, ...changes };
+    
+    /**
+     * REMOVE_LINE_ITEM
+     */
+    case 'REMOVE_LINE_ITEM_PENDING':
+      return { ...state, saving: true };
+
+    case 'REMOVE_LINE_ITEM_REJECTED':
+      changes = {
+        saving: false
+      };
+      return { ...state, ...changes };
+
+    case 'REMOVE_LINE_ITEM_FULFILLED':
       changes = {
         cart: dataFormatter.deserialize(response),
         saving: false
