+++ src/redux/rootReducer.js
@@ -3,7 +3,6 @@ import authReducer from './reducers/authReducer'
 import accountReducer from './reducers/accountReducer'
 import checkoutReducer from './reducers/checkoutReducer'
 import productsReducer from './reducers/productsReducer'
-import cartReducer from './reducers/cartReducer'
 import taxonsReducer from './reducers/taxonsReducer'
 
 const rootReducer = combineReducers({
@@ -11,7 +10,6 @@ const rootReducer = combineReducers({
   account: accountReducer,
   checkout: checkoutReducer,
   products: productsReducer,
-  cart: cartReducer,
   taxons: taxonsReducer
 })
